Move Basecomp to package.
