+++ app/types/types.ts
@@ -1,9 +1,3 @@
-/**
- * Extend multiple classes using mixins.
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type Constructor<T = HTMLElement> = new (...args: any[]) => T;
-
 export interface RestaurantData {
   [region: string]: {
     [city: string]: RestaurantItem[];
