+++ package/src/index.ts
@@ -1,3 +1,4 @@
+import { Basecomp } from "./libs/basecomp/basecomp.ts";
 import { Link } from "./components/link/link.ts";
 import { Params } from "./classes/params/params.ts";
 import { Pathname } from "./classes/pathname/pathname.ts";
@@ -5,4 +6,4 @@ import { PathnameChanged } from "./classes/pathname-changed/pathname-changed.ts"
 import { Route } from "./components/route/route.ts";
 import { Switch } from "./components/switch/switch.ts";
 
-export { Link, Params, Pathname, PathnameChanged, Route, Switch };
+export { Basecomp, Link, Params, Pathname, PathnameChanged, Route, Switch };
